+++ internal/codeintel/dependencies/init.go
@@ -5,26 +5,16 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/codeintel/dependencies/internal/store"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/goroutine"
-	"github.com/sourcegraph/sourcegraph/internal/memo"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
 )
 
-// GetService creates or returns an already-initialized dependencies service.
-// If the service is not yet initialized, it will use the provided dependencies.
-func GetService(db database.DB) *Service {
-	svc, _ := initServiceMemo.Init(serviceDependencies{db})
-	return svc
+func NewService(db database.DB) *Service {
+	return newService(store.New(db, scopedContext("store")), scopedContext("service"))
 }
 
 type serviceDependencies struct {
 	db database.DB
 }
-
-var initServiceMemo = memo.NewMemoizedConstructorWithArg(func(deps serviceDependencies) (*Service, error) {
-	svc := newService(store.New(deps.db, scopedContext("store")), scopedContext("service"))
-
-	return svc, nil
-})
 
 // TestService creates a new dependencies service with noop observation contexts.
 func TestService(db database.DB, gitserver GitserverClient) *Service {